Tweaks to peak detect function

+++ T9_Pedal_Bundle.h
@@ -19,7 +19,7 @@
 void T9PB_begin(void);
 void T9PB_disconnect_all_effects(void);
 void T9PB_hp_volume(float volume);
-int16_t T9PB_peak_detect(int source);
+float T9PB_peak_detect(int source);
 
 
 ///////////////////////////////////////
